Combine Login and Signup into Auth component
