+++ backend/app/routes/auth.py
@@ -17,35 +17,27 @@ bp = Blueprint('auth', __name__, url_prefix='/api/auth')
 @bp.route('/signup', methods=['POST'])
 def signup():
     data = request.get_json()
-    for key, value in data.items():
-        if not value:
-            data[key] = None
+    # for key, value in data.items():
+    #     if not value:
+    #         data[key] = None
     email = data['email']
-    username = data['username']
-    image_url = request.json.get('image_url', None)
+    username = request.json.get('username', None)
+    image_url = 'https://img.icons8.com/doodle/148/000000/test-account.png'
     if User.query.filter_by(email=email).first():
         return {'error': f'The user with email {email} already exists'}, 409
     if username and User.query.filter_by(username=username).first():
         return {'error': f'The user with username {username} already exists'}, 409
-    if not image_url:
-        image_url = 'https://img.icons8.com/doodle/148/000000/test-account.png'
     email_verified = False
     profile_addr = str(uuid4())[:13]
 
     user = User(username=username,
                 first_name=data['first_name'],
-                last_name=data['last_name'],
                 password=data['password'],
                 email=email,
                 email_verified=email_verified,
                 image_url=image_url,
-                user_role=data['user_role'],
-                address=data['address'],
-                lgt=data['lgt'],
-                lat=data['lat'],
                 state=data['state'],
                 city=data['city'],
-                zip_code=data['zip_code'],
                 profile_addr=profile_addr)
 
     db.session.add(user)
