+++ frontend/src/hooks/useRequest.js
@@ -44,7 +44,7 @@ const useRequest = () => {
       const accessToken = localStorage.getItem("access_token");
       const refreshToken = localStorage.getItem("refresh_token");
       if (!accessToken && !refreshToken) {
-        history.push("/signup");
+        history.push("/login");
         return;
       }
     }
